+++ pkg/compactor/compactor_metrics_test.go
@@ -119,6 +119,17 @@ func TestSyncerMetrics(t *testing.T) {
 			cortex_compactor_remaining_planned_compactions{user="aaa"} 377740
 			cortex_compactor_remaining_planned_compactions{user="bbb"} 388850
 			cortex_compactor_remaining_planned_compactions{user="ccc"} 399960
+			# HELP cortex_compactor_compaction_error_total Total number of errors from compactions.
+			# TYPE cortex_compactor_compaction_error_total counter
+			cortex_compactor_compaction_error_total{type="halt",user="aaa"} 444400
+			cortex_compactor_compaction_error_total{type="halt",user="bbb"} 455510
+			cortex_compactor_compaction_error_total{type="halt",user="ccc"} 466620
+			cortex_compactor_compaction_error_total{type="retriable",user="aaa"} 411070
+			cortex_compactor_compaction_error_total{type="retriable",user="bbb"} 422180
+			cortex_compactor_compaction_error_total{type="retriable",user="ccc"} 433290
+			cortex_compactor_compaction_error_total{type="unauthorized",user="aaa"} 477730
+			cortex_compactor_compaction_error_total{type="unauthorized",user="bbb"} 488840
+			cortex_compactor_compaction_error_total{type="unauthorized",user="ccc"} 499950
 	`))
 	require.NoError(t, err)
 
@@ -163,4 +174,13 @@ func generateTestData(cm *compactorMetrics, base float64) {
 	cm.remainingPlannedCompactions.WithLabelValues("aaa").Add(34 * base)
 	cm.remainingPlannedCompactions.WithLabelValues("bbb").Add(35 * base)
 	cm.remainingPlannedCompactions.WithLabelValues("ccc").Add(36 * base)
+	cm.compactionErrorsCount.WithLabelValues("aaa", retriableError).Add(37 * base)
+	cm.compactionErrorsCount.WithLabelValues("bbb", retriableError).Add(38 * base)
+	cm.compactionErrorsCount.WithLabelValues("ccc", retriableError).Add(39 * base)
+	cm.compactionErrorsCount.WithLabelValues("aaa", haltError).Add(40 * base)
+	cm.compactionErrorsCount.WithLabelValues("bbb", haltError).Add(41 * base)
+	cm.compactionErrorsCount.WithLabelValues("ccc", haltError).Add(42 * base)
+	cm.compactionErrorsCount.WithLabelValues("aaa", unauthorizedError).Add(43 * base)
+	cm.compactionErrorsCount.WithLabelValues("bbb", unauthorizedError).Add(44 * base)
+	cm.compactionErrorsCount.WithLabelValues("ccc", unauthorizedError).Add(45 * base)
 }